Move ports to new packages (outbound and inbound)
