+++ src/main/kotlin/no/nb/mlt/wls/application/restApi/ProductController.kt
@@ -6,7 +6,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import io.swagger.v3.oas.annotations.tags.Tag
-import no.nb.mlt.wls.domain.drivingPorts.AddNewItem
+import no.nb.mlt.wls.domain.ports.inbound.AddNewItem
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.PostMapping
