+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/repository/ItemMongoRepository.kt
@@ -2,7 +2,7 @@ package no.nb.mlt.wls.infrastructure.repository
 
 import no.nb.mlt.wls.core.data.HostName
 import no.nb.mlt.wls.domain.Item
-import no.nb.mlt.wls.domain.drivenPorts.ItemRepository
+import no.nb.mlt.wls.domain.ports.outbound.ItemRepository
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
 import org.springframework.stereotype.Component
 import org.springframework.stereotype.Repository
